refactor: MoviesListCellViewModel 기능 MoviesListViewModel로 이동 후 삭제
