+++ BoxOffice/Infra/Util/Observable.swift
@@ -16,4 +16,3 @@ class Observable<T> {
         listener = closure
     }
 }
-
